+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Adjust typegen formatting for metadata documentation
+
+
 ## 6.0.2 Sep 18, 2021
 
 Upgrade priority: Low. Recommended for all users on 6.0.1, especially those using `api.query.substrate.*` calls